[uboot] Fixed compiler warnings and change mode of some files. [Mantis   ID] N/A [Bug    Desc] N/A [Modify Desc] 1. Fixed compiler warnings:                  a. unused variable                  b. printf format               2. Changed mode of some files to 644. [Test   Case] OK [Compile] OK [Coverity] N/A [Selftest] N/A [GPL-licensed & Third-party Source] N/A

+++ common/dlmalloc.c
@@ -1270,7 +1270,7 @@ Void_t* mALLOc(bytes) size_t bytes;
 #else
   if ((long)bytes <= 0) {
     printf("\n\n**** WARN: forbid malloc(0) : return_address=0x%x ****\n\n",
-      __builtin_return_address(0));
+      (unsigned int)__builtin_return_address(0));
     return NULL;
   }
 #endif
